# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Federated Learning API",
        "description": "API documentation for the Medical AI Federated Learning platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/doctor/patients": {
            "get": {
                "tags": [
                    "Doctor — Patients"
                ],
                "summary": "List patients belonging to this doctor's organization",
                "operationId": "2ba1e60abd2443cc11233910cbef238c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "her2",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "er_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "stage_num",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3,
                                4
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of patients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PatientObject"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctor — Patients"
                ],
                "summary": "Register a new patient",
                "operationId": "eca55ee5c95869c1176679022873252b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_identifier",
                                    "er_status",
                                    "pr_status",
                                    "her2_binary",
                                    "age",
                                    "stage_num"
                                ],
                                "properties": {
                                    "patient_identifier": {
                                        "type": "string"
                                    },
                                    "er_status": {
                                        "type": "boolean"
                                    },
                                    "pr_status": {
                                        "type": "boolean"
                                    },
                                    "her2_binary": {
                                        "type": "boolean"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "stage_num": {
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    },
                                    "er_status_missing": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "pr_status_missing": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "fraction_genome_altered": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "buffa_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "ragnum_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "winter_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctor/patients/{id}": {
            "get": {
                "tags": [
                    "Doctor — Patients"
                ],
                "summary": "Show a patient with their full examination & prediction history",
                "operationId": "932c7b969d8d5fe1992c028ce0e82492",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to access this patient"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Doctor — Patients"
                ],
                "summary": "Update clinical data for a patient",
                "operationId": "fd9d9702c4780fd809d15273e9ee874c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "er_status": {
                                        "type": "boolean"
                                    },
                                    "pr_status": {
                                        "type": "boolean"
                                    },
                                    "her2_binary": {
                                        "type": "boolean"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "stage_num": {
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    },
                                    "er_status_missing": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "pr_status_missing": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "fraction_genome_altered": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "buffa_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "ragnum_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "winter_hypoxia_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Doctor — Patients"
                ],
                "summary": "Delete a patient record (only if they have no predictions)",
                "operationId": "c74c2f288b1d159566273837153fc7a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient deleted"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Cannot delete a patient with prediction records"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PatientObject": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "organization_id": {
                        "type": "integer"
                    },
                    "patient_identifier": {
                        "type": "string"
                    },
                    "er_status": {
                        "type": "boolean"
                    },
                    "pr_status": {
                        "type": "boolean"
                    },
                    "her2_binary": {
                        "type": "boolean"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "stage_num": {
                        "type": "integer"
                    },
                    "er_status_missing": {
                        "type": "boolean"
                    },
                    "pr_status_missing": {
                        "type": "boolean"
                    },
                    "fraction_genome_altered": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "buffa_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "ragnum_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "winter_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "HttpOnly cookie issued by Sanctum after a successful `/api/verify-otp` call.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}